requirement.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-bootstrap3==9.1.0
django-cors-headers==2.2.0
django-filter==1.1.0
django-forms-bootstrap==3.1.0
djangorestframework==3.7.7
idna==2.6
image==1.5.20
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2018.3
qrcode==5.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
